src/requirements.txt
django==1.11
django-cleanup==1.0.0
django-datatables-view==1.14.0
django-phonenumber-field==1.3.0
django-sendfile==0.3.11
django-widget-tweaks==1.4.1
pillow==4.2.1
